Update Dockerfile
Mark said, "The error seems to be a Docker error rather than from Virtuoso. I wonder if changing the base image to a newer version of Ubuntu will help. I think Trusty is now EOL which could mean apt is spitting out errors that aren’t making it to the console (and apt failing would be consistent with an i/o error)"

So we're changing the first line from Trusty to Kinetic Ubuntu